+++ src/entrypoints/side.content/components/highlight/index.tsx
@@ -1,14 +1,12 @@
 import type { HighlightData } from '@/entrypoints/side.content/atoms'
+import { useMount } from 'ahooks'
 import getXPath from 'get-xpath'
 import { useAtom } from 'jotai'
 import { Highlighter, Trash2 } from 'lucide-react'
 import { useEffect, useState } from 'react'
 import { highlightsAtom } from '@/entrypoints/side.content/atoms'
 import { cn } from '@/utils/tailwind'
-
-export function generateHighlightId(): string {
-  return `highlight_${Date.now()}`
-}
+import { checkHighlightConflicts, createHighlightData, generateHighlightId, removeAllHighlights, removeHighlight, restoreHighlightFromRange, restoreHighlights, scrollToHighlight } from '../../utils/highlight'
 
 // Color options for highlighting
 const COLOR_OPTIONS = [
@@ -17,206 +15,22 @@ const COLOR_OPTIONS = [
   { color: 'transparent', name: 'Disabled', meaning: 'no_highlight' },
 ]
 
-function getNodeByXPath(xpath: string): Node | null {
-  try {
-    const result = document.evaluate(
-      xpath,
-      document,
-      null,
-      XPathResult.FIRST_ORDERED_NODE_TYPE,
-      null,
-    )
-    return result.singleNodeValue
-  }
-  catch (error) {
-    console.error('XPath查找失败:', error, xpath)
-    return null
-  }
-}
-
-function restoreHighlights(highlights: HighlightData[]) {
-  highlights.forEach((highlight) => {
-    try {
-      const startNode = getNodeByXPath(highlight.startContainer.xpath)
-      const endNode = getNodeByXPath(highlight.endContainer.xpath)
-
-      if (!startNode || !endNode) {
-        console.warn('无法找到高亮的起始或结束节点:', highlight)
-        return
-      }
-
-      const range = document.createRange()
-
-      range.setStart(startNode, highlight.startContainer.offset)
-      range.setEnd(endNode, highlight.endContainer.offset)
-
-      restoreHighlightFromRange(range, highlight)
-    }
-    catch (error) {
-      console.error('恢复高亮失败:', error, highlight)
-    }
-  })
-}
-
-// 从范围恢复高亮
-function restoreHighlightFromRange(range: Range, highlightData: HighlightData) {
-  if (range.startContainer === range.endContainer) {
-    const highlightElement = createHighlightElement(
-      highlightData.id,
-      highlightData.color,
-      highlightData.textContent,
-    )
-    range.surroundContents(highlightElement)
-  }
-  else {
-    // 跨元素高亮
-    const walker = document.createTreeWalker(
-      range.commonAncestorContainer,
-      NodeFilter.SHOW_TEXT,
-      {
-        acceptNode(node) {
-          return range.intersectsNode(node) ? NodeFilter.FILTER_ACCEPT : NodeFilter.FILTER_REJECT
-        },
-      },
-    )
-
-    const textNodes: Node[] = []
-    let currentNode = walker.nextNode()
-    while (currentNode) {
-      textNodes.push(currentNode)
-      currentNode = walker.nextNode()
-    }
-
-    textNodes.forEach((textNode, index) => {
-      let nodeStartOffset = 0
-      let nodeEndOffset = textNode.textContent?.length || 0
-
-      if (textNode === range.startContainer) {
-        nodeStartOffset = range.startOffset
-      }
-      if (textNode === range.endContainer) {
-        nodeEndOffset = range.endOffset
-      }
-
-      if (nodeStartOffset < nodeEndOffset) {
-        const segmentRange = document.createRange()
-        segmentRange.setStart(textNode, nodeStartOffset)
-        segmentRange.setEnd(textNode, nodeEndOffset)
-
-        const highlightElement = createHighlightElement(
-          highlightData.id,
-          highlightData.color,
-          segmentRange.toString(),
-          index,
-        )
-        segmentRange.surroundContents(highlightElement)
-      }
-    })
-  }
-}
-
-function createHighlightElement(id: string, color: string, selectedText: string, index?: number) {
-  const highlightElement = document.createElement('span')
-  highlightElement.style.backgroundColor = color
-  highlightElement.style.cursor = 'pointer'
-  highlightElement.textContent = selectedText
-  highlightElement.setAttribute('data-highlight-id', id)
-  highlightElement.setAttribute('data-segment-index', index?.toString() || '')
-
-  return highlightElement
-}
-
-interface ConflictCheckResult {
-  hasConflict: boolean
-  reason?: string
-  conflictElement?: Element
-}
-
-export function checkHighlightConflicts(range: Range): ConflictCheckResult {
-  // 检查选择范围内是否已包含高亮元素
-  const container = range.commonAncestorContainer instanceof Element
-    ? range.commonAncestorContainer
-    : range.commonAncestorContainer.parentElement || document
-
-  const existingHighlights = container.querySelectorAll('[data-highlight-id]') as NodeListOf<Element>
-
-  for (const highlight of existingHighlights) {
-    // 检查是否在选择范围内
-    if (range.intersectsNode(highlight)) {
-      return {
-        hasConflict: true,
-        reason: '选择范围与现有高亮重叠',
-        conflictElement: highlight,
-      }
-    }
-  }
-
-  // 检查选择范围是否完全在某个高亮内部
-  let currentNode = range.startContainer
-  while (currentNode && currentNode !== document.body) {
-    if (currentNode instanceof Element && currentNode.hasAttribute('data-highlight-id')) {
-      return {
-        hasConflict: true,
-        reason: '选择范围在现有高亮内部',
-        conflictElement: currentNode,
-      }
-    }
-    currentNode = currentNode.parentNode as Node
-  }
-
-  return { hasConflict: false }
-}
-
-interface HighlightProps {
-  className?: string
-}
-
-function Highlight({ className }: HighlightProps) {
+function Highlight() {
   const [highlights, setHighlights] = useAtom(highlightsAtom)
   const [isActive, setIsActive] = useState(true)
   const [highlightColor, setHighlightColor] = useState('#fbbf24') // Default yellow
   const [conflictMessage, setConflictMessage] = useState('')
   const [colorFilter, setColorFilter] = useState<Set<string>>(() => new Set(COLOR_OPTIONS.map(v => v.meaning)))
 
-  useEffect(() => {
+  useMount(() => {
     restoreHighlights(highlights)
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
+  })
 
   const addHighlight = (highlight: HighlightData) => {
     setHighlights(prev => [...prev, highlight])
   }
 
-  const removeHighlight = (id: string) => {
-    setHighlights(prev => prev.filter(h => h.id !== id))
-    // Also remove from DOM
-    const elements = document.querySelectorAll(`[data-highlight-id="${id}"]`)
-    elements.forEach((element) => {
-      const parent = element.parentNode
-      if (parent) {
-        parent.insertBefore(document.createTextNode(element.textContent || ''), element)
-        parent.removeChild(element)
-        parent.normalize()
-      }
-    })
-  }
-
-  const removeAllHighlights = () => {
-    // Remove from DOM
-    highlights.forEach((highlight) => {
-      const elements = document.querySelectorAll(`[data-highlight-id="${highlight.id}"]`)
-      elements.forEach((element) => {
-        const parent = element.parentNode
-        if (parent) {
-          parent.insertBefore(document.createTextNode(element.textContent || ''), element)
-          parent.removeChild(element)
-          parent.normalize()
-        }
-      })
-    })
-  }
-
-  const createHighlight = (range: Range, textContent: string) => {
+  const createHighlight = (range: Range) => {
     // Don't create highlight if disabled
     if (!isActive || highlightColor === 'transparent') {
       return
@@ -230,78 +44,12 @@ function Highlight({ className }: HighlightProps) {
     }
 
     try {
-      const highlightId = generateHighlightId()
-
-      // 获取开始和结束容器的详细信息
-      const startContainer = range.startContainer
-      const endContainer = range.endContainer
-
-      // 构建高亮数据，支持跨元素
-      const highlightData: HighlightData = {
-        id: highlightId,
-        textContent,
-        color: highlightColor,
-        startContainer: {
-          xpath: getXPath(startContainer, { ignoreId: true }),
-          offset: range.startOffset,
-        },
-        endContainer: {
-          xpath: getXPath(endContainer, { ignoreId: true }),
-          offset: range.endOffset,
-        },
-        timestamp: Date.now(),
-        context: '',
-        pageUrl: window.location.origin + window.location.pathname,
-      }
+      const highlightData = createHighlightData(range, highlightColor)
 
       // 保存高亮数据
       addHighlight(highlightData)
 
-      // 创建高亮元素
-      if (range.startContainer === range.endContainer) {
-        const highlightElement = createHighlightElement(highlightId, highlightColor, textContent)
-        range.surroundContents(highlightElement)
-      }
-      else {
-        // 跨元素高亮处理
-        const walker = document.createTreeWalker(
-          range.commonAncestorContainer,
-          NodeFilter.SHOW_TEXT,
-          {
-            acceptNode(node) {
-              return range.intersectsNode(node) ? NodeFilter.FILTER_ACCEPT : NodeFilter.FILTER_REJECT
-            },
-          },
-        )
-
-        const textNodes: Text[] = []
-        let currentNode = walker.nextNode() as Text
-        while (currentNode) {
-          textNodes.push(currentNode)
-          currentNode = walker.nextNode() as Text
-        }
-
-        textNodes.forEach((textNode, index) => {
-          let startOffset = 0
-          let endOffset = textNode.textContent?.length || 0
-
-          if (textNode === range.startContainer) {
-            startOffset = range.startOffset
-          }
-          if (textNode === range.endContainer) {
-            endOffset = range.endOffset
-          }
-
-          if (startOffset < endOffset) {
-            const segmentRange = document.createRange()
-            segmentRange.setStart(textNode, startOffset)
-            segmentRange.setEnd(textNode, endOffset)
-
-            const span = createHighlightElement(highlightId, highlightColor, segmentRange.toString(), index)
-            segmentRange.surroundContents(span)
-          }
-        })
-      }
+      restoreHighlightFromRange(range, highlightData)
     }
     catch (error) {
       console.error('创建高亮失败:', error)
@@ -343,30 +91,6 @@ function Highlight({ className }: HighlightProps) {
     })
   }
 
-  // Scroll to highlight
-  const scrollToHighlight = (highlight: HighlightData) => {
-    const element = document.querySelector(`[data-highlight-id="${highlight.id}"]`)
-    if (element) {
-      element.scrollIntoView({
-        behavior: 'smooth',
-        block: 'center',
-        inline: 'nearest',
-      })
-
-      // Add temporary highlight effect
-      const originalBoxShadow = (element as HTMLElement).style.boxShadow
-      const originalTransition = (element as HTMLElement).style.transition
-
-      ;(element as HTMLElement).style.boxShadow = '0 0 8px rgba(59, 130, 246, 0.5)'
-      ;(element as HTMLElement).style.transition = 'box-shadow 0.3s ease'
-
-      setTimeout(() => {
-        ;(element as HTMLElement).style.boxShadow = originalBoxShadow
-        ;(element as HTMLElement).style.transition = originalTransition
-      }, 2000)
-    }
-  }
-
   useEffect(() => {
     if (!isActive || highlightColor === 'transparent')
       return
@@ -383,7 +107,7 @@ function Highlight({ className }: HighlightProps) {
         return
       }
 
-      createHighlight(range, selectedText)
+      createHighlight(range)
       selection.removeAllRanges()
     }
 
@@ -397,7 +121,7 @@ function Highlight({ className }: HighlightProps) {
   }, [isActive, highlightColor, highlights])
 
   return (
-    <div className={cn('border-b border-border', className)}>
+    <div className={cn('border-b border-border')}>
       {/* Header */}
       <div className="flex w-full items-center justify-between p-3">
         <div className="flex items-center gap-2">
@@ -414,7 +138,7 @@ function Highlight({ className }: HighlightProps) {
             <button
               type="button"
               onClick={() => {
-                removeAllHighlights()
+                removeAllHighlights(highlights)
                 setHighlights([])
               }}
               className="flex items-center gap-1 px-2 py-1 text-xs text-red-600 hover:text-red-700 hover:bg-red-50 rounded transition-colors"
@@ -429,7 +153,7 @@ function Highlight({ className }: HighlightProps) {
               const nextState = !isActive
               setIsActive(nextState)
               if (nextState === false) {
-                removeAllHighlights()
+                removeAllHighlights(highlights)
               }
               else {
                 restoreHighlights(highlights)
@@ -613,7 +337,10 @@ function Highlight({ className }: HighlightProps) {
                             </button>
                             <button
                               type="button"
-                              onClick={() => removeHighlight(highlight.id)}
+                              onClick={() => {
+                                removeHighlight(highlight.id)
+                                setHighlights(prev => prev.filter(h => h.id !== highlight.id))
+                              }}
                               className="ml-2 p-1 text-red-500 hover:text-red-700 hover:bg-red-50 rounded opacity-0 group-hover:opacity-100 transition-opacity"
                               title="Remove highlight"
                             >
